# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Notion KB Manager API
  description: |
    Production-ready API for managing knowledge base content with Notion integration.

    ## Features
    - 🔗 Link management and validation
    - 📄 Content parsing and extraction
    - 🤖 AI-powered content processing
    - 📝 Notion database integration
    - 💾 Backup and restore
    - ⚙️ Configuration management
    - 📊 Task monitoring and reporting

    ## Authentication
    Most endpoints require JWT authentication. Include your token in the `Authorization` header:
    ```
    Authorization: Bearer <your_token>
    ```

    To obtain a token, use the `/api/auth/login` endpoint with valid credentials.

    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - Default: 1000 requests/hour, 100 requests/minute
    - Config endpoints: 100 requests/hour
    - Parsing: 10 requests/minute
    - AI processing: 20 requests/minute
    - Backups: 5 requests/hour

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Request limit
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Reset timestamp

  version: 1.0.0
  contact:
    name: Notion KB Manager Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://staging-api.example.com/api
    description: Staging server
...
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /backup/:
    post:
      tags: [Backup]
      summary: Create backup
      description: Create database and file backup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [manual, auto]
                  default: manual
                include_database:
                  type: boolean
                  default: true
                include_files:
                  type: boolean
                  default: true
                retention_days:
                  type: integer
      responses:
        '200':
          description: Backup created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      backup_id:
                        type: integer
                      filename:
                        type: string
                      size:
                        type: integer
                        description: Backup size in bytes
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

    get:
      tags: [Backup]
      summary: List backups
      description: Get list of available backups
      responses:
        '200':
          description: Backup list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Backup'
  /backup/{backup_id}:
    get:
      tags: [Backup]
      summary: Get backup details
      parameters:
        - $ref: '#/components/parameters/BackupId'
      responses:
        '200':
          description: Backup details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Backup'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Backup]
      summary: Delete backup
      parameters:
        - $ref: '#/components/parameters/BackupId'
      responses:
        '200':
          description: Backup deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /backup/{backup_id}/restore:
    post:
      tags: [Backup]
      summary: Restore from backup
      description: Restore database and files from backup
      parameters:
        - $ref: '#/components/parameters/BackupId'
      responses:
        '200':
          description: Backup restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== Configuration ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  parameters:
...
    BackupId:
      name: backup_id
      in: path
      required: true
      schema:
        type: integer
      description: Backup ID
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
...
    Backup:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        type:
          type: string
          enum: [manual, auto]
        size:
          type: integer
        created_at:
          type: string
          format: date-time
        retention_days:
          type: integer
...
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_001
              message: Authorization required
              details: Please provide a valid access token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RES_001
              message: Resource not found
              details: Link with ID 123 not found
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_001
              message: Rate limit exceeded
              details: Too many requests. Please try again later.
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: SYS_001
              message: Internal server error
              details: An unexpected error occurred